motors/motor_requirements.txt
pyserial
git+https://github.com/liambindle/pyvesc.git

